douyu: update
close #448
